chore(deps): package version upgrades
