+++ src/app.module.ts
@@ -5,8 +5,7 @@ import { ConfigModule } from "@nestjs/config";
 import { ThrottlerModule, ThrottlerGuard } from "@nestjs/throttler";
 import { LoggerModule } from "nestjs-pino";
 import { default as pinoPretty } from "pino-pretty";
-import { default as env } from "@/common/env";
-import { validationSchema } from "@/common/env.validation";
+import { default as env, validationSchema } from "@/common/env";
 import { ApiController, PortalsController } from "@/controllers";
 import { ApiService, CronService, PortalsService } from "@/services";
