+++ src/common/env.ts
@@ -1,3 +1,4 @@
+import * as Joi from "joi";
 import { CommonConstants } from "@/common/enums";
 
 export default () => ({
@@ -10,4 +11,12 @@ export default () => ({
   SWAGGER_USE_METADATA_FILE:
     process.env.SWAGGER_USE_METADATA_FILE === CommonConstants.TRUE_STRING,
   GOOGLE_ANALYTICS_TAG: process.env.GOOGLE_ANALYTICS_TAG,
+});
+
+export const validationSchema = Joi.object({
+  NODE_ENV: Joi.string().required().default("development"),
+  THROTTLER_TTL: Joi.number().required().default(60),
+  THROTTLER_REQ_PER_TTL: Joi.number().required().default(600),
+  SWAGGER_USE_METADATA_FILE: Joi.string().required().default(true),
+  GOOGLE_ANALYTICS_TAG: Joi.string().required(),
 });
