+++ src/common/index.ts
@@ -3,7 +3,6 @@ export * from "./constants";
 export * from "./enums";
 export * from "./env";
 export * from "./formatting";
-export * from "./functions";
 export * from "./routes";
 export * from "./types";
 export * from "./validation";
