+++ src/configs/scrape-zagreb.config.ts
@@ -1,3 +1,4 @@
+import { randomUUID } from "crypto";
 import { Portals } from "@/common/enums";
 import { ScraperConfig } from "@/common/types";
 
@@ -8,11 +9,11 @@ export const ScrapeZagrebConfig: ScraperConfig = {
   icon: "https://www.zagreb.info/wp-content/uploads/2018/02/zagreb-favicon-2.png",
   rss: false,
   roots: [
-    "https://www.zagreb.info/category/hrvatska/page/1/",
-    "https://www.zagreb.info/category/hrvatska/page/2/",
-    "https://www.zagreb.info/category/hrvatska/page/3/",
-    "https://www.zagreb.info/category/hrvatska/page/4/",
-    "https://www.zagreb.info/category/hrvatska/page/5/",
+    "https://www.zagreb.info/category/vijesti/page/1/",
+    "https://www.zagreb.info/category/vijesti/page/2/",
+    "https://www.zagreb.info/category/vijesti/page/3/",
+    "https://www.zagreb.info/category/vijesti/page/4/",
+    "https://www.zagreb.info/category/vijesti/page/5/",
     "https://www.zagreb.info/category/sport/page/1/",
     "https://www.zagreb.info/category/sport/page/2/",
     "https://www.zagreb.info/category/sport/page/3/",
@@ -20,13 +21,8 @@ export const ScrapeZagrebConfig: ScraperConfig = {
     "https://www.zagreb.info/category/sport/page/5/",
   ],
   linker: "div.item-details > h3 > a",
-  id: (link: string) =>
-    link
-      .substring(link.lastIndexOf("-") + 1)
-      .replace("/", "-")
-      .replace("/", "")
-      .split("-")[1],
-  remove1: ["img", "iframe", "figure", "div.wpipa-container"],
+  id: (link: string) => randomUUID(),
+  remove1: ["img", "iframe", "figure", "div.wpipa-container", "div.fc-ab-root"],
   title: {
     find: "h1.entry-title",
   },
